feat(any): use mandatory type and instantiate features in aspect
