--- src/main/java/fr/epita/assistants/myide/domain/entity/features/Clean.java
@@ -1,7 +1,0 @@
-package fr.epita.assistants.myide.domain.entity.features;
-
-public
-class Clean extends Maven_features {
-public
-  Clean() { super(FeaturesTypes.Maven.CLEAN, "mvn clean"); }
-}
