--- src/main/java/fr/epita/assistants/myide/domain/entity/features/Compile.java
@@ -1,7 +1,0 @@
-package fr.epita.assistants.myide.domain.entity.features;
-
-public
-class Compile extends Maven_features {
-public
-  Compile() { super(FeaturesTypes.Maven.COMPILE, "mvn compile"); }
-}
