--- src/main/java/fr/epita/assistants/myide/domain/entity/features/Exec.java
@@ -1,7 +1,0 @@
-package fr.epita.assistants.myide.domain.entity.features;
-
-public
-class Exec extends Maven_features {
-public
-  Exec() { super(FeaturesTypes.Maven.EXEC, "mvn exec"); }
-}
