--- src/main/java/fr/epita/assistants/myide/domain/entity/features/Install.java
@@ -1,7 +1,0 @@
-package fr.epita.assistants.myide.domain.entity.features;
-
-public
-class Install extends Maven_features {
-public
-  Install() { super(FeaturesTypes.Maven.INSTALL, "mvn install"); }
-}
