--- src/main/java/fr/epita/assistants/myide/domain/entity/features/Test.java
@@ -1,7 +1,0 @@
-package fr.epita.assistants.myide.domain.entity.features;
-
-public
-class Test extends Maven_features {
-public
-  Test() { super(FeaturesTypes.Maven.TEST, "mvn test"); }
-}
