--- src/main/java/fr/epita/assistants/myide/domain/entity/features/Tree.java
@@ -1,7 +1,0 @@
-package fr.epita.assistants.myide.domain.entity.features;
-
-public
-class Tree extends Maven_features {
-public
-  Tree() { super(FeaturesTypes.Maven.TREE, "mvn dependency:tree"); }
-}
